Replace schema editor drawer with sidebar pane
- Add DashboardCodePane component for sidebar-based code editing
- Add 'code' to DashboardSidebarPaneName type
- Auto-resize pane to 500px when code editor opens
- Remove SchemaV2EditorDrawer overlay pattern
- Remove openV2SchemaEditor() method from DashboardScene
- Update EditSchemaV2Button to use editPane.openPane('code')
