+++ eslint-suppressions.json
@@ -1637,6 +1637,11 @@
       "count": 2
     }
   },
+  "public/app/features/dashboard-scene/edit-pane/DashboardEditPaneSplitter.tsx": {
+    "react-hooks/rules-of-hooks": {
+      "count": 5
+    }
+  },
   "public/app/features/dashboard-scene/inspect/HelpWizard/HelpWizard.tsx": {
     "no-restricted-syntax": {
       "count": 3
