+++ public/app/features/dashboard-scene/edit-pane/DashboardEditPaneRenderer.tsx
@@ -14,6 +14,7 @@ import { DashboardFiltersOverviewPane } from '../scene/dashboard-filters-overvie
 import { ToolbarActionProps } from '../scene/new-toolbar/types';
 import { dynamicDashNavActions } from '../utils/registerDynamicDashNavAction';
 
+import { DashboardCodePane } from './DashboardCodePane';
 import { DashboardEditPane } from './DashboardEditPane';
 import { ShareExportDashboardButton } from './DashboardExportButton';
 import { DashboardOutline } from './DashboardOutline';
@@ -89,6 +90,11 @@ export function DashboardEditPaneRenderer({ editPane, dashboard }: Props) {
           />
         </Sidebar.OpenPane>
       )}
+      {openPane === 'code' && (
+        <Sidebar.OpenPane>
+          <DashboardCodePane dashboard={dashboard} />
+        </Sidebar.OpenPane>
+      )}
       <Sidebar.Toolbar>
         {isEditing && (
           <>
@@ -130,6 +136,13 @@ export function DashboardEditPaneRenderer({ editPane, dashboard }: Props) {
                 'dashboard-scene.dashboard-edit-pane-renderer.title-feedback-dashboard-editing-experience',
                 'Give feedback on the new dashboard editing experience'
               )}
+            />
+            <Sidebar.Button
+              tooltip={t('dashboard.sidebar.edit-schema.tooltip', 'Edit as code')}
+              title={t('dashboard.sidebar.edit-schema.title', 'Code')}
+              icon="brackets-curly"
+              onClick={() => editPane.openPane('code')}
+              active={openPane === 'code'}
             />
             {config.featureToggles.dashboardUndoRedo && (
               <>
